// src/client/src/app/guards/auth.ts
/**
 * variable dependecies for communication of express to angular
 */
export interface Auth {
  PieData: number;
  pieData: number;
  PieReason: string;
  pieReason: string;
  title: string;
  message: string;
  success: boolean;
  token: string;
  user: string;
  list: string;
  errorToken: boolean;
  usersCount: number;
}